Revert "Merge remote-tracking branch 'origin/amos-branch' into carlos-branch"
This reverts commit deb89c5d3f0580b63fda512306d341b943a873f0, reversing
changes made to c3dc7138955a36cc2d9f0f06a45785b248819f0d.
