+++ app/src/main/AndroidManifest.xml
@@ -41,10 +41,10 @@
             android:name=".recetas_guardadas_activity"
             android:exported="false" />
         <activity
-            android:name=".ui.editar_informacion_perfil_activity"
+            android:name=".editar_informacion_perfil_activity"
             android:exported="false" />
         <activity
-            android:name=".ui.PerfilConfigActivity"
+            android:name=".PerfilConfigActivity"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
