+++ app/src/main/java/flores/pablo/sazonforaneo/PerfilConfigActivity.kt
@@ -1,0 +1,18 @@
+package flores.pablo.sazonforaneo
+
+import android.os.Bundle
+import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
+
+class PerfilConfigActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContentView(R.layout.activity_perfil_config)
+
+
+
+    }
+}
