+++ app/src/main/java/flores/pablo/sazonforaneo/ui/ExplorarActivity.kt
@@ -14,24 +14,11 @@ class ExplorarActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_explorar)
 
-        val fragmentToShow = intent.getStringExtra("fragment_to_show")
+        replaceFragment(CategoriasFragment())
 
         val bottomNav = findViewById<BottomNavigationView>(R.id.bottom_navigation_view)
 
-        when(fragmentToShow) {
-            "mis_recetas" -> {
-                replaceFragment(MisRecetasFragment())
-                bottomNav.selectedItemId = R.id.nav_mis_recetas
-            }
-            "explorar" -> {
-                replaceFragment(ExplorarFragment())
-                bottomNav.selectedItemId = R.id.nav_explorar
-            }
-            else -> {
-                replaceFragment(CategoriasFragment())
-                bottomNav.selectedItemId = R.id.nav_categorias
-            }
-        }
+        bottomNav.selectedItemId = R.id.nav_categorias
 
         bottomNav.setOnItemSelectedListener { item ->
             when (item.itemId) {
@@ -51,7 +38,6 @@ class ExplorarActivity : AppCompatActivity() {
             }
         }
     }
-
 
     private fun replaceFragment(fragment: Fragment) {
         supportFragmentManager.beginTransaction()
