--- app/src/main/java/flores/pablo/sazonforaneo/ui/Usuario.kt
@@ -1,10 +1,0 @@
-package flores.pablo.sazonforaneo.ui
-
-data class Usuario (
-    var nombre: String,
-    var correo: String,
-    var telefono: String,
-    var fechaNacimiento: String
-
-)
-
