--- app/src/main/res/drawable/etiqueta_redonda.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape android:shape="rectangle" xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <solid android:color="@color/baby_green"/>
-    <corners android:radius="30dp"/>
-
-</shape>
