+++ app/src/main/res/layout/activity_perfil_config.xml
@@ -6,7 +6,7 @@
     android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.PerfilConfigActivity">
+    tools:context=".PerfilConfigActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
