--- app/src/main/res/layout/etiqueta_element.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="60dp"
-    android:orientation="vertical"
-    android:padding="5dp"
-    android:layout_height="30dp">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@drawable/etiqueta_redonda"
-        android:orientation="vertical"
-        android:gravity="center"
-        android:elevation="5dp">
-
-        <TextView
-            android:id="@+id/nombreEtiqueta"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Pizza"
-            android:textSize="14sp"
-            android:textColor="@color/black"
-            android:textStyle="bold"/>
-
-
-
-    </LinearLayout>
-
-</LinearLayout>
