+++ app/src/main/res/values/colors.xml
@@ -9,6 +9,5 @@
     <color name="baby_gray">#F6F6F6</color>
     <color name="medium_gray">#CFCFCF</color>
     <color name="wine">#44291D</color>
-    <color name="baby_green">#6FE04D</color>
     
 </resources>
